main/app/migrations/requirements.txt
psycopg2-binary>=2.9.6
configparser>=5.3.0
tabulate>=0.9.0 